fix `huge` rdma_recv_block_type no effect issue.
It has no effect for `huge` type of recv_block_size
which is as large as 2^21, because the `block_size`
variable is defined as type `uint16_t`, its maximum
value is 2^16. So, change the type of `block_size`
to `uint32_t` with relevant updates.

+++ src/brpc/rdma/rdma_endpoint.h
@@ -216,7 +216,7 @@ private:
     // Data address of _rbuf
     std::vector<void*> _rbuf_data;
     // Remote block size for receiving
-    uint16_t _remote_recv_block_size;
+    uint32_t _remote_recv_block_size;
 
     // The number of new recv WRs acked to the remote side
     uint16_t _accumulated_ack;
